Finish setup in Simulation.cpp

+++ sources/users/Researcher.cpp
@@ -1,7 +1,7 @@
 
 #include "Researcher.h"
 
-Researcher::Researcher(struct Group group, int grant) :
+Researcher::Researcher(struct Group group, double grant) :
         User(group.expoParameter, group.budget, group.permission),
         grant(grant)
     {
